refactor: 기능별 컴포넌트 나누기
