+++ src/app/exercise-details/components/ExerciseInfo.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const ExerciseInfo = () => {
+  return (
+    <div className="mx-auto px-5">
+      <h3 className="font-bold text-lg leading-[26px] text-text-main mb-1 mt-3">
+        덤벨 라잉 외회전 숄더 로테이션
+      </h3>
+      <p className="text-sm leading-5 text-text-light mb-4">가슴 : 스미스머신</p>
+      <div className="w-full h-[184px] bg-backgrounds-light mb-4"></div>
+    </div>
+  );
+};
+
+export default ExerciseInfo;
